requirements.txt
torch
torchvision
torchaudio
torchsummary
fastapi
uvicorn
streamlit
pillow
python-multipart
onnxruntime
numpy==1.26

